Support for variants of USB Ser/HID (e.g TI MSP430 HID API) added, report format can be specified in the USB Ser/HID settings (feature requests #193 and #209)  > HID implementation improved to properly handle vendor-defined usages  > Communication with TI MSP-EXP430F5529LP partly successful, but writing to stream makes it crash (IOException)  > Command line args pending

+++ trunk/MKY/MKY.IO.Usb/Device.cs
@@ -165,7 +165,7 @@ namespace MKY.IO.Usb
 		{
 			// Set the size property of attributes to the number of bytes in the structure.
 			Win32.Hid.NativeTypes.HIDD_ATTRIBUTES attributes = new Win32.Hid.NativeTypes.HIDD_ATTRIBUTES();
-			attributes.Size = Marshal.SizeOf(attributes);
+			attributes.Size = (UInt32)Marshal.SizeOf(attributes);
 
 			if (Win32.Hid.NativeMethods.HidD_GetAttributes(deviceHandle, ref attributes))
 			{
